IS13 add items to database

+++ app/platforms/dnd5/item.rb
@@ -4,7 +4,7 @@ module Dnd5
   class ItemData
     include StoreModel::Model
 
-    attribute :weight, :integer
+    attribute :weight, :float
     attribute :price, :integer
   end
 end
